requirements.txt
Flask==2.1.2
Werkzeug==2.1.2
requests==2.27.1
beautifulsoup4==4.11.1
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
scikit-image==0.25.2
opencv-python==4.11.0.86
scikit-learn==1.6.1
tqdm==4.67.1
imagehash==4.3.2
torch==2.6.0
torchvision==0.21.0
plotly